# api/userProgram/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a swag test http server.",
        "title": "USER WEB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7001",
    "basePath": "/api/v1",
    "paths": {
...
        "/mer/addresses": {
            "post": {
                "description": "获取商户地址信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取商户地址信息",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/merchantService.QueryMerchantAddressInfoListModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/merchantService.MerchantAddressInfoModel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/mer/list/search": {
            "post": {
                "description": "获取商户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取商户列表",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/merchantService.GetMerchantListModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/merchantService.MerchantListModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/mer/product/get/list": {
            "get": {
                "description": "获取商品列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取商品列表",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.GetMerProductInfoListModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/productService.ProductSmallInfoModel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/mer/shopQRCode/get": {
            "post": {
                "description": "描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取商户店铺二维码",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.GetMerchantShopQRCodeModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.GetMerchantShopQRCodeModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
        "/merchant/get": {
            "get": {
                "description": "获取商户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取商户信息",
                "parameters": [
                    {
                        "description": "body参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/merchantService.GetMerchantInfoModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"返回信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/merchantService.MerInfoModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "401": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    },
                    "500": {
                        "description": "错误",
                        "schema": {
                            "$ref": "#/definitions/controller.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.GetMerProductInfoListModel": {
            "type": "object",
            "properties": {
                "merCode": {
                    "type": "string"
                },
                "merId": {
                    "type": "integer"
                }
            }
        },
        "controller.GetMerchantShopQRCodeModel": {
            "type": "object",
            "properties": {
                "merCode": {
                    "type": "string"
                },
                "merId": {
                    "type": "integer"
                }
            }
        },
...
        "controller.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "merchantService.GetMerchantInfoModel": {
            "type": "object",
            "properties": {
                "isWhereUserId": {
                    "type": "boolean"
                },
                "merCode": {
                    "type": "string"
                },
                "merId": {
                    "type": "integer"
                }
            }
        },
        "merchantService.GetMerchantListModel": {
            "type": "object",
            "properties": {
                "computeDistance": {
                    "type": "boolean"
                },
                "distance": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "sortType": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "merchantService.MerInfoModel": {
            "type": "object",
            "required": [
                "resources"
            ],
            "properties": {
                "activeComment": {
                    "type": "boolean"
                },
                "activeGoods": {
                    "type": "boolean"
                },
                "activeService": {
                    "type": "boolean"
                },
                "address": {
                    "type": "MerAddressModel"
                },
                "businessTime": {
                    "type": "MerBusinessTime"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "logo": {
                    "type": "MerResourceModel"
                },
                "name": {
                    "type": "string"
                },
                "phones": {
                    "type": "string"
                },
                "resources": {
                    "type": "array",
                    "items": {
                        "type": "MerResourceModel"
                    }
                },
                "typeId": {
                    "type": "integer"
                }
            }
        },
        "merchantService.MerchantAddressInfoModel": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "merId": {
                    "type": "integer"
                },
                "merName": {
                    "type": "string"
                },
                "phones": {
                    "type": "string"
                }
            }
        },
        "merchantService.MerchantListModel": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "image": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "merId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "merchantService.QueryMerchantAddressInfoListModel": {
            "type": "object",
            "properties": {
                "merIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "productService.ProductSmallInfoModel": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "object"
                },
                "brandId": {
                    "type": "integer"
                },
                "currentPrice": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "guid": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "keywords": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "maxPrice": {
                    "type": "number"
                },
                "merId": {
                    "type": "integer"
                },
                "merName": {
                    "type": "string"
                },
                "minPrice": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "originalPrice": {
                    "type": "number"
                },
                "productType": {
                    "type": "integer"
                },
                "sales": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                },
                "typeId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}